refactor: fix casing
