+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "Spib",
+  "name": "spib",
   "version": "0.1.0",
   "private": false,
-  "homepage": "https://Spib.no",
+  "homepage": "https://spib.no",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
