+++ src/App.jsx
@@ -1,5 +1,5 @@
 import './App.css';
-import { FetchAndDisplayImage } from './Fetchimage';
+import { FetchAndDisplayImage } from './FetchImage';
 import { useEffect, useState } from 'react';
 import React from 'react';
 import ReactDOM from 'react-dom'
